+++ src/Plugin/GraphQL/DataProducer/PageListProducer.php
@@ -1,0 +1,128 @@
+<?php
+
+namespace Drupal\thunder_gqls\Plugin\GraphQL\DataProducer;
+
+use Drupal\Core\Cache\RefinableCacheableDependencyInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Drupal\graphql\Plugin\GraphQL\DataProducer\DataProducerPluginBase;
+use Drupal\thunder_gqls\Wrappers\QueryConnection;
+use GraphQL\Error\UserError;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+/**
+ * @DataProducer(
+ *   id = "thunder_page_list_producer",
+ *   name = @Translation("Load entities"),
+ *   description = @Translation("Loads a list of entities of a type."),
+ *   produces = @ContextDefinition("any",
+ *     label = @Translation("Entity connection")
+ *   ),
+ *   consumes = {
+ *     "type" = @ContextDefinition("string",
+ *       label = @Translation("Node Type"),
+ *       required = TRUE
+ *     ),
+ *     "bundles" = @ContextDefinition("any",
+ *       label = @Translation("Entity bundle(s)"),
+ *       multiple = TRUE,
+ *       required = FALSE
+ *     ),
+ *     "offset" = @ContextDefinition("integer",
+ *       label = @Translation("Offset"),
+ *       required = FALSE
+ *     ),
+ *     "limit" = @ContextDefinition("integer",
+ *       label = @Translation("Limit"),
+ *       required = FALSE
+ *     )
+ *   }
+ * )
+ */
+class PageListProducer extends DataProducerPluginBase implements ContainerFactoryPluginInterface {
+
+  const MAX_LIMIT = 100;
+
+  /**
+   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
+   */
+  protected $entityTypeManager;
+
+  /**
+   * {@inheritdoc}
+   *
+   * @codeCoverageIgnore
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('entity_type.manager')
+    );
+  }
+
+  /**
+   * Articles constructor.
+   *
+   * @param array $configuration
+   *   The plugin configuration.
+   * @param string $pluginId
+   *   The plugin id.
+   * @param mixed $pluginDefinition
+   *   The plugin definition.
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entityTypeManager
+   *
+   * @codeCoverageIgnore
+   */
+  public function __construct(
+    array $configuration,
+    $pluginId,
+    $pluginDefinition,
+    EntityTypeManagerInterface $entityTypeManager
+  ) {
+    parent::__construct($configuration, $pluginId, $pluginDefinition);
+    $this->entityTypeManager = $entityTypeManager;
+  }
+
+  /**
+   * @param string $type
+   * @param array|null $bundles
+   *   List of bundles to be filtered.
+   * @param int $offset
+   * @param int $limit
+   * @param \Drupal\Core\Cache\RefinableCacheableDependencyInterface $metadata
+   *
+   * @return \Drupal\thunder_gqls\Wrappers\QueryConnection
+   * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
+   * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
+   */
+  public function resolve(string $type, ?array $bundles = NULL, $offset, $limit, RefinableCacheableDependencyInterface $metadata) {
+    if ($limit > static::MAX_LIMIT) {
+      throw new UserError(sprintf('Exceeded maximum query limit: %s.', static::MAX_LIMIT));
+    }
+
+    $storage = $this->entityTypeManager->getStorage($type);
+    $entityType = $storage->getEntityType();
+    $query = $storage->getQuery()
+      ->currentRevision()
+      ->accessCheck();
+
+    //$query->condition($entityType->getKey('bundle'), $bundles);
+    if (isset($bundles)) {
+      $bundle_key = $entityType->getKey('bundle');
+      if (!$bundle_key) {
+        throw new UserError('No bundles defined for given entity type.');
+      }
+      $query->condition($bundle_key, $bundles, "IN");
+    }
+
+    $query->range($offset, $limit);
+
+    $metadata->addCacheTags($entityType->getListCacheTags());
+    $metadata->addCacheContexts($entityType->getListCacheContexts());
+
+    return new QueryConnection($query);
+  }
+
+}
